% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigWigGenerator.R
\name{create.bigwig.chaitanya}
\alias{create.bigwig.chaitanya}
\title{Create bigwig file from raw genomic profile}
\usage{
create.bigwig.chaitanya(profiles, file, model = NULL, truncate.tol = 1e-04,
  normalize = "genome")
}
\arguments{
\item{profiles}{Raw genomic score profiles, created using \code{fetch.genomic.profiles}}

\item{file}{Name of the bigwig file that is to be created.}

\item{model}{Model used for normalization based on theoretical optimal site (see 'optimize').}

\item{truncate.tol}{Cutoff below which affinities are assumed to be zero.}

\item{normalize}{Whether to use genomic maximum ('genome') or optimal site for model ('model') as reference for relative binding affinity. Note that model needs to be specified when the latter option is used.}

\item{footprint}{Footprint size of the NRLB model that was used to generate the genomic profile.}
}
\value{
None
}
\description{
Create bigwig file from raw genomic profile
}
\note{
First, the affinity scores for the forward and reverse strand are added up for each window. Next, the value assigned to any particular base pair is the maximum over all windows containing that base pair.
}
\examples{

m = NRLBtools::hox.models()$ExdScr
gp = fetch.genomic.profiles("http://bussemakerlab.org/NRLB/dm3/ExdScr", 
                            genome = BSgenome.Dmelanogaster.UCSC.dm3, 
                            model = m, 
                            chr.names = c("chrXHet", "chrYHet"))

my.bwfile = tempfile(fileext=".bw")
create.bigwig.chaitanya(profiles = gp, file = my.bwfile, model = m)

bw = import(my.bwfile)
bw

}
