% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic-profiles.R
\name{fetch.as.track}
\alias{fetch.as.track}
\title{Fetch pre-computed genomic score profiles as GRanges object}
\usage{
fetch.as.track(url.prefix, genome, chr.names = NULL, normalize = "none",
  model = NULL)
}
\arguments{
\item{url.prefix}{Prefix of URL containing precomputed profiles}

\item{genome}{Complete genome sequence in BSgenome format}

\item{chr.names}{Chromosome(s) to which profile should be restricted (optional; the default is to include profiles for all chromosomes in the genome)}

\item{model}{An NRLB model (required to determine footprint)}

\item{rc}{Whether to swap forward and reverse profiles}
}
\value{
A nested list containing forward and reverse NRLB score profiles for each chromosome.
}
\description{
Fetch pre-computed genomic score profiles as GRanges object
}
\examples{

m = NRLBtools::hox.models()$ExdScr
gr = fetch.as.track("http://bussemakerlab.org/NRLB/dm3/ExdScr", 
                    genome = BSgenome.Dmelanogaster.UCSC.dm3, 
                    model = m, normalize = "none",
                    chr.names = c("chrXHet", "chrYHet")
                    )
gr

v = Views(BSgenome.Dmelanogaster.UCSC.dm3$chrXHet, start=1:6, width=footprint.size(m))
sapply(v, function(x) score.genome(x, model=m))

}
