% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-parser.R
\name{fetch.genomic.profiles}
\alias{fetch.genomic.profiles}
\title{Fetch pre-computed genomic score profiles}
\usage{
fetch.genomic.profiles(url.prefix, genome, model)
}
\arguments{
\item{url.prefix}{Prefix of URL containing precomputed profiles}

\item{genome}{Complete genome sequence in BSgenome format}

\item{model}{An NRLB model (required to determine footprint)}
}
\value{
A nested list containing forward and reverse NRLB score profiles for each chromosome.
}
\description{
Fetch pre-computed genomic score profiles
}
\examples{

m = NRLBtools::hox.models()$ExdScr
NRLBtools::fetch.genomic.profiles()

}
