% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-parser.R
\name{file.parser}
\alias{file.parser}
\title{Parse NRLB model file}
\usage{
file.parser(fileName, fitInfoOnly = FALSE)
}
\arguments{
\item{fileName}{name of NRLB model fit in CSV format}

\item{fitInfoOnly}{}
}
\value{
an NRLB fit object
}
\description{
Parse NRLB model file
}
\examples{

}
