% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-parser.R
\name{footprint.size}
\alias{footprint.size}
\title{Footprint size associated with model}
\usage{
footprint.size(model)
}
\arguments{
\item{model}{NRLB model}
}
\value{
Footprint size in bp
}
\description{
Footprint size associated with model
}
\examples{

m = NRLBtools::hox.models()$ExdScr
footprint.size(m)

}
