% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic-profiles.R
\name{genomic.max}
\alias{genomic.max}
\title{Maximum NRLB score in genomic profile}
\usage{
genomic.max(profiles)
}
\arguments{
\item{profiles}{Raw genomic score profiles, created using \code{fetch.genomic.profiles}}
}
\value{
Maximumover all chromosomes and both binding directions
}
\description{
Maximum NRLB score in genomic profile
}
\examples{

}
