% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hox-models.R
\name{hox.models}
\alias{hox.models}
\title{Collection of NRLB models for Hox complexes}
\usage{
hox.models()
}
\value{
Nested list containing model data and metadata
}
\description{
Collection of NRLB models for Hox complexes
}
\examples{

models = hox.models()
names(models)

}
