% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-models.R
\name{load.models}
\alias{load.models}
\title{Load an NRLB Model File}
\usage{
load.models(fileName)
}
\arguments{
\item{fileName}{file path to the NRLB fit results file.}
}
\value{
An NRLB model object.
}
\description{
A function used to load the results of an NRLB run (models) in CSV format into R and return an NRLB model object. This object can be used by the \code{model.info},
\code{logo}, and \code{score.seq} functions.
}
\examples{
m = NRLBtools::load.models(fileName = system.file("extdata", "MAX-NRLBConfig.csv", package = "NRLBtools"))

}
