% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogoGenerator.R
\name{LogoGenerator}
\alias{LogoGenerator}
\title{Use LogoGenerator tool from REDUCE Suite}
\usage{
LogoGenerator(output, file, logo, title, ymin, ymax)
}
\arguments{
\item{output}{Directory in which output files are stored. The default is \code{tempdir()}.}
}
\value{
An optional error code
}
\description{
Use LogoGenerator tool from REDUCE Suite
}
