% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogoGenerator_cmdline}
\alias{LogoGenerator_cmdline}
\title{Run 'LogoGenerator' tool from REDUCE Suite}
\usage{
LogoGenerator_cmdline(argstr)
}
\arguments{
\item{argstr}{Commandline argument string}
}
\value{
None [return path to logo file in future?]
}
\description{
Run 'LogoGenerator' tool from REDUCE Suite
}
