% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max-seq.R
\name{max.seq}
\alias{max.seq}
\title{Find the highest affinity sequence for a model}
\usage{
\method{max}{seq}(fits, index, mode = NULL)
}
\arguments{
\item{fits}{TO_BE_ADDED}

\item{index}{TO_BE_ADDED}

\item{mode}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Find the highest affinity sequence for a model
}
\note{
uses dynamic programming
}
\examples{

}
