% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-info.R
\name{model.info}
\alias{model.info}
\title{Display model information}
\usage{
model.info(models = NULL, index = NULL, nrlb.model = NULL)
}
\arguments{
\item{models}{an NRLB model object.}

\item{index}{numeric indicating the index of the model in the NRLB model object.}

\item{nrlb.model}{numeric index of the pre-built NRLB model.}
}
\value{
a data frame.
}
\description{
A function used to display various levels of information for NRLB model objects and pre-built NRLB models.
}
\details{
\code{model.info} provides many different levels of information depending on the number of arguments passed to it. When run with
no arguments, model.info returns a data frame containing the information of all the NRLB models stored in the package and used in the Rastogi, et al.
(2018) paper. 

When the index of an NRLB model (\code{nrlb.model}) is provided, a data frame containing the name, footprint size (\code{k}), mode relative affinity
(\code{RelativeAffinity}), and the optimal sequence (\code{BestSequence}) is provided for all modes in the model. The nonspecific binding mode is 
denoted by \code{NSB} and only has an associated relative affinity. 

When only an NRLB model object (\code{models}) is provided, a data frame containing summary information for the models in the model object is returned.

When both an NRLB model object and a model index is provided, a data frame containing the name, footprint size (\code{k}), mode relative affinity
(\code{RelativeAffinity}), and the optimal sequence (\code{BestSequence}) is provided for all modes in the model. The nonspecific binding mode is 
denoted by \code{NSB} and only has an associated relative affinity.
}
\examples{
model.info()

model.info(nrlb.model = 15)

m = NRLBtools::load.models(fileName = system.file("extdata", "MAX-NRLBConfig.csv", package = "NRLBtools"))
model.info(m)

model.info(m, 8)

}
