% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-score-genome.R
\name{nrlb.plot.score.genome}
\alias{nrlb.plot.score.genome}
\title{Plot affinities of a sequence}
\usage{
nrlb.plot.score.genome(genomicSequence, fits, index, mode = NULL,
  rc = FALSE, nPeaks = NULL, annotate = FALSE, rescale = NULL,
  genomicSequence2 = NULL)
}
\arguments{
\item{genomicSequence}{TO_BE_ADDED}

\item{fits}{TO_BE_ADDED}

\item{index}{TO_BE_ADDED}

\item{mode}{TO_BE_ADDED}

\item{rc}{TO_BE_ADDED}

\item{nPeaks}{TO_BE_ADDED}

\item{annotate}{TO_BE_ADDED}

\item{rescale}{TO_BE_ADDED}

\item{genomicSequence2}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Plot affinities of a sequence
}
\examples{

}
