% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-NRLBtools.R
\docType{package}
\name{NRLBtools}
\alias{NRLBtools}
\title{NRLBtools: utitilies for working with NRLB models}
\description{
The NRLB package provides tools for working with protein-DNA recognition models created by the No Read Left Behind (NRLB) algorithm developed by the Bussemaker Lab in the Department of Biological Sciences at Columbia University.
}
\author{
Chaitanya Rastogi and Harmen J. Bussemaker
}
