% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{ql}
\alias{ql}
\title{Quick Logo}
\usage{
ql(fits, index, mode = NULL, rc = FALSE)
}
\arguments{
\item{fits}{NRLB fit set}

\item{index}{Index of model within fit set}

\item{mode}{Mode within model}

\item{rc}{Whether to use model in reverse-complement form}
}
\value{
Plots an image representing an energy logo for the model
}
\description{
Quick Logo
}
\examples{

m = NRLBtools::hox.models()$ExdSrc
ql(m$fits, m$index, m$mode)

}
