% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-genome.R
\name{score.genome}
\alias{score.genome}
\title{Score genome with a fit}
\usage{
score.genome(genomicSequence, fits, index, mode = NULL, rc = FALSE)
}
\arguments{
\item{genomicSequence}{TO_BE_ADDED}

\item{fits}{TO_BE_ADDED}

\item{index}{TO_BE_ADDED}

\item{mode}{TO_BE_ADDED}

\item{rc}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Score genome with a fit
}
\examples{

}
