% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic-profiles.R
\name{score.parser}
\alias{score.parser}
\title{Read in binary files containing genomic scoring results}
\usage{
score.parser(genome, score.dir, prot.name, chr.name, k)
}
\arguments{
\item{genome}{TO_BE_ADDED}

\item{score.dir}{TO_BE_ADDED}

\item{prot.name}{TO_BE_ADDED}

\item{chr.name}{TO_BE_ADDED}

\item{k}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Read in binary files containing genomic scoring results
}
\examples{

}
