% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TestParseStringVectorArgs}
\alias{TestParseStringVectorArgs}
\title{Run 'LogoGenerator' tool from REDUCE Suite}
\usage{
TestParseStringVectorArgs(arguments)
}
\arguments{
\item{argv}{A commandline option string}
}
\value{
An optional error code
}
\description{
Run 'LogoGenerator' tool from REDUCE Suite
}
