% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigWigGenerator.R
\name{create.bigwig}
\alias{create.bigwig}
\title{Read in binary files containing genomic scoring results}
\usage{
create.bigwig(profiles, file, model, truncate.tol = 1e-04,
  normalize = "model")
}
\arguments{
\item{profiles}{Raw genomic score profiles, created using \code{fetch.genomic.profiles}}

\item{file}{Name of the bigwig file that is to be created.}

\item{truncate.tol}{Cutoff below which affinities are assumed to be zero.}

\item{footprint}{Footprint size of the NRLB model that was used to generate the genomic profile.}
}
\value{
None
}
\description{
Read in binary files containing genomic scoring results
}
\examples{

m = NRLBtools::hox.models()$ExdScr
gp = fetch.genomic.profiles("http://bussemakerlab.org/NRLB/dm3/ExdScr", 
                            genome = BSgenome.Dmelanogaster.UCSC.dm3, 
                            model = m, 
                            chr.names = c("chrXHet", "chrYHet"))

create.bigwig(profiles = gp, file = tempfile(fileext=".bw"), model = m)

}
