% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic-profiles.R
\name{fetch.genomic.profiles}
\alias{fetch.genomic.profiles}
\title{Fetch pre-computed genomic score profiles as nested list object}
\usage{
fetch.genomic.profiles(url.prefix, genome, model, chr.names = NULL)
}
\arguments{
\item{url.prefix}{Prefix of URL containing precomputed profiles}

\item{genome}{Complete genome sequence in BSgenome format}

\item{model}{An NRLB model (required to determine footprint)}

\item{chr.names}{Chromosome(s) to which profile should be restricted (optional; the default is to include profiles for all chromosomes in the genome)}
}
\value{
A nested list containing forward and reverse NRLB score profiles for each chromosome.
}
\description{
Fetch pre-computed genomic score profiles as nested list object
}
\examples{

m = NRLBtools::hox.models()$ExdScr
gp = fetch.genomic.profiles("http://bussemakerlab.org/NRLB/dm3/ExdScr", 
                            BSgenome.Dmelanogaster.UCSC.dm3, m, 
                            c("chrXHet", "chrYHet"))
names(gp)
names(gp$chrXHet)


head(gp$chrXHet$fwd)
head(gp$chrXHet$rev)

v = Views(BSgenome.Dmelanogaster.UCSC.dm3$chrXHet, start=1:6, width=footprint.size(m))
sapply(v, function(x) score.genome(x, model=m))

}
