% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo}
\alias{logo}
\title{Generate PSAM}
\usage{
logo(fits, index = NULL, mode = NULL, rc = FALSE, display = TRUE,
  betas = NULL, save.path = NULL, isPDF = TRUE, title = NULL,
  ylim = NULL, fit.name = NULL, l.pad = 0, r.pad = 0, l.del = 0,
  r.del = 0)
}
\arguments{
\item{fits}{TO_BE_ADDED}

\item{index}{TO_BE_ADDED}

\item{mode}{TO_BE_ADDED}

\item{rc}{TO_BE_ADDED}

\item{display}{TO_BE_ADDED}

\item{betas}{TO_BE_ADDED}

\item{save.path}{TO_BE_ADDED}

\item{isPDF}{TO_BE_ADDED}

\item{title}{TO_BE_ADDED}

\item{ylim}{TO_BE_ADDED}

\item{fit.name}{TO_BE_ADDED}

\item{l.pad}{TO_BE_ADDED}

\item{r.pad}{TO_BE_ADDED}

\item{l.del}{TO_BE_ADDED}

\item{r.del}{TO_BE_ADDED}
}
\value{
TO_BE_ADDED
}
\description{
Generate PSAM
}
\examples{

}
